backend/app/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
slowapi
yt-dlp
python-dotenv